requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
streamlit>=1.31.0
bsvlib>=0.10.0
tqdm>=4.66.0
